# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for TodoList Application",
        "title": "Todo App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/register_user": {
            "post": {
                "description": "Регистрация нового пользователя с вводом обязательных и необязательных данных.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "register-user",
                "parameters": [
                    {
                        "description": "Информация о пользователе. Время в формате 1990-01-01T00:00:00Z",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.RequestRegister"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Не проходит валидация пользователя"
                    }
                }
            }
        },
        "/update_user": {
            "put": {
                "description": "Обновление полей пользователя с вводом обязательных и необязательных данных.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Обновление полей пользователя",
                "operationId": "update-user",
                "parameters": [
                    {
                        "description": "Токен и информация, которую пользователь хочет изменить. Время в формате 1990-01-01T00:00:00Z",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.RequestUpdate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Данные пользователя обновлен",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Неверный токен"
                    },
                    "403": {
                        "description": "Нельзя сменить логин или email"
                    }
                }
            }
        },
        "/user_login": {
            "post": {
                "description": "Вход для пользователя с login и password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Вход для пользователя",
                "operationId": "user-login",
                "parameters": [
                    {
                        "description": "Информация для входа. Время в формате 1990-01-01T00:00:00Z",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь авторизован",
                        "schema": {
                            "type": "token"
                        }
                    },
                    "400": {
                        "description": "Логин не найден"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "storage.LoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "storage.RequestRegister": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "storage.RequestUpdate": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}